accurately calculate screen real estate

+++ compleet-client/src/bindings/api/window.rs
@@ -14,7 +14,7 @@ pub fn win_close(lua: &Lua, winid: u32, force: bool) -> LuaResult<()> {
 }
 
 /// Binding to `vim.api.nvim_win_get_cursor`
-pub fn win_get_cursor(lua: &Lua, winid: u32) -> LuaResult<(u32, u32)> {
+pub fn win_get_cursor(lua: &Lua, winid: u16) -> LuaResult<(u16, u16)> {
     let position = super::api(lua)?
         .get::<&str, LuaFunction>("nvim_win_get_cursor")?
         .call::<_, Table>(winid)?;
@@ -43,14 +43,14 @@ pub fn win_get_position(lua: &Lua, winid: u32) -> LuaResult<(u16, u16)> {
 }
 
 /// Binding to `vim.api.nvim_win_get_width`
-pub fn win_get_width(lua: &Lua, winid: u32) -> LuaResult<u32> {
+pub fn win_get_width(lua: &Lua, winid: u16) -> LuaResult<u16> {
     super::api(lua)?
         .get::<&str, LuaFunction>("nvim_win_get_width")?
         .call(winid)
 }
 
 /// Binding to `vim.api.nvim_win_get_width`
-pub fn win_get_height(lua: &Lua, winid: u32) -> LuaResult<u32> {
+pub fn win_get_height(lua: &Lua, winid: u16) -> LuaResult<u16> {
     super::api(lua)?
         .get::<&str, LuaFunction>("nvim_win_get_height")?
         .call(winid)
